requirements.txt
# Core dependencies
click>=8.0.0
pyyaml>=6.0
requests>=2.28.0
packaging>=21.0
rich>=12.0.0
jinja2>=3.0.0

# SBOM format libraries
cyclonedx-bom>=4.0.0
spdx-tools>=0.8.0
packageurl-python>=0.11.0

# Security scanning
safety>=2.3.0
pip-audit>=2.4.0

# Language-specific parsers
toml>=0.10.2
defusedxml>=0.7.1
lxml>=4.9.0

# Async and parallel processing
aiohttp>=3.8.0

# Report generation
matplotlib>=3.5.0
pandas>=1.4.0
tabulate>=0.9.0

# Version control
gitpython>=3.1.0

